Bug #128: Ensure that client is computing before issuing cancel
This test is trying to verify that the cancel token was working
as expected. However there was a race condition in the test that
meant if the cancel was acted on before the computeAsync was run
the computeAsync would not be run. This caused an intermittent
test failure because cancellationHappened[0] would never be set
to true, even though the client would act on the cancellation.

Signed-off-by: Jonah Graham <jonah@kichwacoders.com>