Eclipse Feature contains incorrect license (#563)
The lsp4j-feature still contains EPL v1.0 even though lsp4j is
now licensed under EPL v2.0. I noticed this because it is the
license from the feature that is shown to the user when they
are installing it into their IDE -- it was clearly incorrect.

This change makes the lsp4j-feature take advantage of the CBI
shared license functionality to generate and populate the license
information automatically at build time. The exact license to
use is specified in the target platform as the current latest
version of EPL v2.0.

Signed-off-by: Mat Booth <mat.booth@gmail.com>