Don't annotate SymbolInformation as Deprecated
While SymbolInformation class is deprecated in the
LSP specification. It is not annotated with Deprecated
annotation because the TextDocumentService.documentSymbol
method requires it and causes deprecated warning in
the code of all users of that API.

Instead we mark all the methods/constructors as
@Deprecated so that consumers that are actually using
instances of SymbolInformation get deprecated warnings,
while avoiding deprecated warnings for just referring
to the type.

Fixes #697