Void return types need to always be null value
Void type in Java cannot be instantiated, therefore if a
type is declared as Void the fromJson should always set that
value to null.

If you want to have a generic return type that allows non-null
and null results, use a type other than Void as the return type.

Fixes #721