Use/allow latest versions of dependencies for guava and jakarta.websocket
A new version of Guava has been released (33) that LSP4J works with and
websocket is now published in a "better" way (direct from maven) so we
republish that version now.

Consumers shouldn't need to change their dependencies as the lower bounds
are unchanged.

This applies the changes in #785 and subsequent commits to main branch.