Remove javadoc links to non-public fields
This is a place where we meet the problem of Java vs Typescript.

In Typescript, the fields are directly accessible, and matches
what the LSP spec says. But in Java we wrap the fields in a
get/set pair, meaning the field is not visible and generates
a javadoc warning.