Disable value of method parameter is unused warnings in Java
There are lots of these warnings and they can be ignored with this
setting, or `@param params` can be added to Javadoc block for each of them.

As this warning was recently introduced, and is not a common default,
I have elected to turn the warning off. To turn this warning on again,
ensure that all existing cases are resolved.