Disable potential resource leak warnings in Java
The Gson library that LSP4J is heavily dependent on uses the errorprone
annotations. If Eclipse supported warning suppression based on the
errorprone annotations, the dozens of potential resource leak warnings
in the LSP4J code would be automatically suppressed.

This happens because Gson classes such as JsonWriter returns `this`
in many of its methods, and since JsonWriter is a Closeable, the
JDT warns about it. See for example JsonWrite.nullValue. Gson annotates
this method as CanIgnoreReturnValue, meaning no potential resource leak.

If JDT gets support for error prone annotations, this warning can be
re-enabled.